+++ src-docs/src/views/list_group/list_group_links.js
@@ -1,0 +1,97 @@
+import React, { Component, Fragment } from 'react';
+
+import {
+  EuiListGroup,
+  EuiSpacer,
+  EuiSwitch,
+  EuiCode,
+  EuiFlexGroup,
+  EuiFlexItem,
+} from '../../../../src/components';
+
+const myContent = [
+  {
+    label: 'First link',
+    href: '/#/display/list-group',
+    iconType: 'calendar',
+    size: 's',
+  },
+  {
+    label: 'Second link is active',
+    href: '/#/display/list-group',
+    isActive: true,
+    iconType: 'clock',
+    size: 's',
+  },
+  {
+    label: 'Third link is disabled',
+    href: '/#/display/list-group',
+    isDisabled: true,
+    iconType: 'compute',
+    size: 's',
+  },
+  {
+    label: 'Fourth link',
+    href: '/#/display/list-group',
+    iconType: 'copyClipboard',
+    size: 's',
+  },
+  {
+    label: 'Fifth link',
+    href: '/#/display/list-group',
+    iconType: 'crosshairs',
+    size: 's',
+  },
+];
+
+
+export default class extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      flushWidth: false,
+      showBorder: false,
+    };
+  }
+
+  toggleFlushWidth = () => {
+    this.setState(prevState => ({ flushWidth: !prevState.flushWidth }));
+  };
+
+  toggleBorder = () => {
+    this.setState(prevState => ({ showBorder: !prevState.showBorder }));
+  };
+
+  render() {
+    const {
+      flushWidth,
+      showBorder,
+    } = this.state;
+
+    return (
+      <Fragment>
+        <EuiFlexGroup alignItems="center">
+          <EuiFlexItem grow={false}>
+            <EuiSwitch
+              label={<span>Show as <EuiCode>flush</EuiCode></span>}
+              checked={this.state.flushWidth}
+              onChange={this.toggleFlushWidth}
+            />
+          </EuiFlexItem>
+          <EuiFlexItem grow={false}>
+            <EuiSwitch
+              label={<span>Show as <EuiCode>bordered</EuiCode></span>}
+              checked={this.state.showBorder}
+              onChange={this.toggleBorder}
+            />
+          </EuiFlexItem>
+        </EuiFlexGroup>
+
+        <EuiSpacer size="l" />
+
+        <EuiListGroup flush={flushWidth} bordered={showBorder} listItems={myContent} />
+      </Fragment>
+    );
+  }
+}